Move pkg/model into pkg/config
Sensible semantic grouping.
