+++ CONTRIBUTING.md
@@ -64,9 +64,9 @@ As much as possible, this is attempting to follow the [Standard Go Project Layou
 
 - `cmd/` - The root `cog` command.
 - `pkg/cli/` - CLI commands.
+- `pkg/config` - Everything `cog.yaml` related.
 - `pkg/docker/` - Various interfaces with Docker for building and running containers.
 - `pkg/dockerfile/` - Creates Dockerfiles.
-- `pkg/model/` - Models, versions, and configs (`cog.yaml`).
 - `pkg/predict/` - Runs predictions on models.
 - `pkg/util/` - Various packages that aren't part of Cog. They could reasonably be separate re-usable projects.
 - `python/` - The Cog Python library.
