+++ pkg/config/compatibility.go
@@ -1,4 +1,4 @@
-package model
+package config
 
 import (
 	_ "embed"
