+++ pkg/config/config_test.go
@@ -1,4 +1,4 @@
-package model
+package config
 
 import (
 	"testing"
